Individuals schools pages now have tabs.
These tabs list reviews like the main page review cards. This required
some changing, so a few of the links are funky... a href seems to break
angular material.

The other two tabs are for courses and professors. Since those cards
aren't currently in a tag, it just shows names for now.